implement ptc analysis
